Change phoneNumber's datatype to allow for more digits + Add new script

+++ server/models/helpers/user.js
@@ -6,7 +6,7 @@ export default {
       othername text,
       email text NOT NULL,
       password text NOT NULL,
-      phoneNumber INTEGER,
+      phoneNumber BIGINT,
       username text,
       registered DATE DEFAULT NOW(),
       isAdmin boolean DEFAULT FALSE
